refactor: move all tests to tests/ directory with lib.rs crate structure
Extract inline #[cfg(test)] blocks and submodule test files into
integration tests under tests/. Create lib.rs for public module exports,
simplify main.rs to thin wrapper. All 52 tests pass (+ 1 ignored).
