+++ crates/crabitat-control-plane/src/db/repos.rs
@@ -96,113 +96,3 @@ pub fn update(
         .map_err(|e| e.to_string())? ;
     Ok(affected > 0)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::db;
-
-    fn setup() -> Connection {
-        let conn = Connection::open_in_memory().unwrap();
-        conn.pragma_update(None, "foreign_keys", "ON").unwrap();
-        db::migrate(&conn);
-        conn
-    }
-
-    #[test]
-    fn insert_and_list_repos() {
-        let conn = setup();
-        insert(&conn, "owner", "name", None, None).unwrap();
-        let repos = list(&conn).unwrap();
-        assert_eq!(repos.len(), 1);
-        assert_eq!(repos[0].owner, "owner");
-        assert_eq!(repos[0].name, "name");
-    }
-
-    #[test]
-    fn soft_delete_repo() {
-        let conn = setup();
-        let repo = insert(&conn, "owner", "name", None, None).unwrap();
-
-        // Before delete: list should have 1
-        assert_eq!(list(&conn).unwrap().len(), 1);
-
-        // Delete it
-        assert!(delete(&conn, &repo.repo_id).unwrap());
-
-        // After delete: list should be empty
-        assert_eq!(list(&conn).unwrap().len(), 0);
-
-        // But get_by_id should still return it for history
-        let deleted_repo = get_by_id(&conn, &repo.repo_id).unwrap().unwrap();
-        assert!(deleted_repo.deleted_at.is_some());
-    }
-
-    #[test]
-    fn delete_idempotency() {
-        let conn = setup();
-        let repo = insert(&conn, "owner", "name", None, None).unwrap();
-
-        // First delete returns true
-        assert!(delete(&conn, &repo.repo_id).unwrap());
-        // Second delete returns false (already deleted)
-        assert!(!delete(&conn, &repo.repo_id).unwrap());
-    }
-
-    #[test]
-    fn get_nonexistent_repo() {
-        let conn = setup();
-        assert!(get_by_id(&conn, "no-such-id").unwrap().is_none());
-    }
-
-    #[test]
-    fn re_add_soft_deleted_repo() {
-        let conn = setup();
-        let repo = insert(&conn, "owner", "name", None, None).unwrap();
-        delete(&conn, &repo.repo_id).unwrap();
-
-        // After partial index fix, re-adding after soft-delete should succeed
-        let result = insert(&conn, "owner", "name", None, None);
-        assert!(
-            result.is_ok(),
-            "Should be able to re-add soft-deleted repo, but got: {:?}",
-            result.err()
-        );
-    }
-
-    #[test]
-    fn multiple_soft_delete_re_add_cycles() {
-        let conn = setup();
-
-        // Cycle 1: insert -> delete
-        let repo1 = insert(&conn, "owner", "name", None, None).unwrap();
-        delete(&conn, &repo1.repo_id).unwrap();
-
-        // Cycle 2: insert -> delete
-        let repo2 = insert(&conn, "owner", "name", None, None).unwrap();
-        assert_ne!(repo1.repo_id, repo2.repo_id);
-        delete(&conn, &repo2.repo_id).unwrap();
-
-        // Cycle 3: insert (leave active)
-        let repo3 = insert(&conn, "owner", "name", None, None).unwrap();
-        assert_ne!(repo2.repo_id, repo3.repo_id);
-
-        // Only the latest active repo should appear in list
-        let repos = list(&conn).unwrap();
-        assert_eq!(repos.len(), 1);
-        assert_eq!(repos[0].repo_id, repo3.repo_id);
-    }
-
-    #[test]
-    fn active_duplicate_repo_rejected() {
-        let conn = setup();
-        insert(&conn, "owner", "name", None, None).unwrap();
-        let result = insert(&conn, "owner", "name", None, None);
-        assert!(
-            result.is_err(),
-            "Should NOT be able to add duplicate active repo"
-        );
-        let err = result.err().unwrap();
-        assert!(err.contains("UNIQUE constraint failed"), "got: {err}");
-    }
-}
