+++ crates/crabitat-control-plane/src/db/workflows.rs
@@ -117,159 +117,3 @@ pub fn update_flavor(
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::db;
-
-    fn setup() -> Connection {
-        let conn = Connection::open_in_memory().unwrap();
-        conn.pragma_update(None, "foreign_keys", "ON").unwrap();
-        db::migrate(&conn);
-        conn
-    }
-
-    #[test]
-    fn insert_and_list_flavors() {
-        let conn = setup();
-        let f = insert_flavor(
-            &conn,
-            "my-workflow",
-            "rust",
-            &["rust/base.md".into(), "rust/extra.md".into()],
-        )
-        .unwrap();
-        assert_eq!(f.workflow_name, "my-workflow");
-        assert_eq!(f.name, "rust");
-        assert_eq!(f.prompt_paths, vec!["rust/base.md", "rust/extra.md"]);
-
-        let flavors = list_flavors_for_workflow(&conn, "my-workflow").unwrap();
-        assert_eq!(flavors.len(), 1);
-        assert_eq!(flavors[0].name, "rust");
-    }
-
-    #[test]
-    fn count_flavors() {
-        let conn = setup();
-        assert_eq!(count_flavors_for_workflow(&conn, "wf").unwrap(), 0);
-        insert_flavor(&conn, "wf", "a", &[]).unwrap();
-        insert_flavor(&conn, "wf", "b", &["x.md".into()]).unwrap();
-        assert_eq!(count_flavors_for_workflow(&conn, "wf").unwrap(), 2);
-    }
-
-    #[test]
-    fn delete_existing_flavor() {
-        let conn = setup();
-        let f = insert_flavor(&conn, "wf", "rust", &[]).unwrap();
-        assert!(delete_flavor(&conn, &f.flavor_id).unwrap());
-        assert_eq!(list_flavors_for_workflow(&conn, "wf").unwrap().len(), 0);
-    }
-
-    #[test]
-    fn delete_nonexistent_flavor() {
-        let conn = setup();
-        assert!(!delete_flavor(&conn, "no-such-id").unwrap());
-    }
-
-    #[test]
-    fn duplicate_flavor_name_rejected() {
-        let conn = setup();
-        insert_flavor(&conn, "wf", "rust", &[]).unwrap();
-        let err = insert_flavor(&conn, "wf", "rust", &[]).unwrap_err();
-        assert!(err.contains("already exists"), "got: {err}");
-    }
-
-    #[test]
-    fn same_flavor_name_different_workflows() {
-        let conn = setup();
-        insert_flavor(&conn, "wf-a", "rust", &[]).unwrap();
-        insert_flavor(&conn, "wf-b", "rust", &[]).unwrap();
-        assert_eq!(count_flavors_for_workflow(&conn, "wf-a").unwrap(), 1);
-        assert_eq!(count_flavors_for_workflow(&conn, "wf-b").unwrap(), 1);
-    }
-
-    #[test]
-    fn list_flavors_empty_workflow() {
-        let conn = setup();
-        let flavors = list_flavors_for_workflow(&conn, "nonexistent").unwrap();
-        assert!(flavors.is_empty());
-    }
-
-    #[test]
-    fn flavor_prompt_paths_roundtrip() {
-        let conn = setup();
-        let paths = vec![
-            "a/b.md".to_string(),
-            "c/d.md".to_string(),
-            "e.md".to_string(),
-        ];
-        let f = insert_flavor(&conn, "wf", "multi", &paths).unwrap();
-        assert_eq!(f.prompt_paths, paths);
-
-        let listed = list_flavors_for_workflow(&conn, "wf").unwrap();
-        assert_eq!(listed[0].prompt_paths, paths);
-    }
-
-    #[test]
-    fn flavors_sorted_by_name() {
-        let conn = setup();
-        insert_flavor(&conn, "wf", "zulu", &[]).unwrap();
-        insert_flavor(&conn, "wf", "alpha", &[]).unwrap();
-        insert_flavor(&conn, "wf", "mike", &[]).unwrap();
-        let flavors = list_flavors_for_workflow(&conn, "wf").unwrap();
-        let names: Vec<&str> = flavors.iter().map(|f| f.name.as_str()).collect();
-        assert_eq!(names, vec!["alpha", "mike", "zulu"]);
-    }
-
-    #[test]
-    fn re_add_soft_deleted_flavor() {
-        let conn = setup();
-        let f = insert_flavor(&conn, "wf", "rust", &[]).unwrap();
-        delete_flavor(&conn, &f.flavor_id).unwrap();
-
-        // After partial index fix, re-adding after soft-delete should succeed
-        let result = insert_flavor(&conn, "wf", "rust", &[]);
-        assert!(
-            result.is_ok(),
-            "Should be able to re-add soft-deleted flavor, but got: {:?}",
-            result.err()
-        );
-    }
-
-    #[test]
-    fn multiple_soft_delete_re_add_cycles() {
-        let conn = setup();
-
-        // Cycle 1: insert -> delete
-        let f1 = insert_flavor(&conn, "wf", "rust", &[]).unwrap();
-        delete_flavor(&conn, &f1.flavor_id).unwrap();
-
-        // Cycle 2: insert -> delete
-        let f2 = insert_flavor(&conn, "wf", "rust", &[]).unwrap();
-        assert_ne!(f1.flavor_id, f2.flavor_id);
-        delete_flavor(&conn, &f2.flavor_id).unwrap();
-
-        // Cycle 3: insert (leave active)
-        let f3 = insert_flavor(&conn, "wf", "rust", &[]).unwrap();
-        assert_ne!(f2.flavor_id, f3.flavor_id);
-
-        // Only the latest active flavor should appear in list
-        let flavors = list_flavors_for_workflow(&conn, "wf").unwrap();
-        assert_eq!(flavors.len(), 1);
-        assert_eq!(flavors[0].flavor_id, f3.flavor_id);
-    }
-
-    #[test]
-    fn active_duplicate_flavor_rejected() {
-        let conn = setup();
-        insert_flavor(&conn, "wf", "rust", &[]).unwrap();
-        let result = insert_flavor(&conn, "wf", "rust", &[]);
-        assert!(
-            result.is_err(),
-            "Should NOT be able to add duplicate active flavor"
-        );
-        let err = result.err().unwrap();
-        assert!(err.contains("already exists"), "got: {err}");
-    }
-}
